backend/pyproject.toml
[project]
name = "maki-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "ffmpeg-python>=0.2.0",
    "groq>=0.31.1",
    "numpy>=2.3.3",
    "openai-whisper",
    "pyannote-audio>=3.4.0",
    "pyannote-core>=5.0.0",
    "pymongo>=4.15.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "speechbrain>=1.0.3",
    "torch>=2.8.0",
    "uvicorn[standard]>=0.37.0",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
pyannote-audio = { git = "https://github.com/pyannote/pyannote-audio.git" }
